+++ coral/media/js/views/components/plugins/open-issue-report-workflow.js
@@ -31,11 +31,11 @@ const openWorkflowViewModel = function (params) {
   };
 
   this.getParentTileOptions = async (resourceId) => {
-    const tiles = await this.fetchTileData(resourceId, '20017860-d711-11ee-9dd0-0242ac120006');
+    const tiles = await this.fetchTileData(resourceId, '3497eaa2-fa47-59fa-a890-035b7c211278');
     this.parentTileOptions(
       tiles.map((tile, idx) => {
         return {
-          text: tile?.data['2001a33a-d711-11ee-9dd0-0242ac120006']?.en?.value,
+          text: tile?.data['b075893b-848d-5520-9d52-3c3dfbecde16']?.en?.value,
           tile: tile,
           id: tile.parenttile
         };
@@ -57,7 +57,7 @@ const openWorkflowViewModel = function (params) {
   });
 
   this.selectedIssueReport.subscribe((tileId) => {
-    this.addtionalConfigData()['parentTileIds']['d3ff3fe6-d62b-11ee-9454-0242ac180006'] = tileId;
+    this.addtionalConfigData()['parentTileIds']['7f835acf-5601-5dae-ac0f-6f030fc50ee7'] = tileId;
     this.setAdditionalOpenConfigData();
   });
 };